速度控制改为浮点数
提高速度和加强减速，还没试；加了热量限制，还没试

+++ APP/RemoteTask.h
@@ -87,9 +87,9 @@ typedef enum
 
 typedef __packed struct
 {
-    int16_t forward_back_ref;
-    int16_t left_right_ref;
-    int16_t rotate_ref;
+    float forward_back_ref;
+    float left_right_ref;
+    float rotate_ref;
 }ChassisSpeed_Ref_t;
 
 typedef __packed struct
